Revert "2.16.11-SNAPSHOT"
This reverts commit 4fd6fc8b51362349218f62c8aa649b9b6d58cd67.
