+++ bybitlinear-api/src/main/java/io/contek/invoker/bybitlinear/api/rest/user/UserRestApi.java
@@ -44,8 +44,8 @@ public final class UserRestApi {
     return new GetWalletBalance(actor, context);
   }
 
-  public PostPositionSwitchIsolated postPositionSwitchIsolated() {
-    return new PostPositionSwitchIsolated(actor, context);
+  public PostLeverageSave postLeverageSave() {
+    return new PostLeverageSave(actor, context);
   }
 
   public PostOrderCancel postOrderCancel() {
